+++ examples/flow-react-example/src/ShadowedValuesInModules.gen.js
@@ -1,0 +1,13 @@
+/** 
+ * @flow strict
+ * @generated
+ * @nolint
+ */
+/* eslint-disable */
+
+// $FlowExpectedError: Reason checked type sufficiently
+import * as ShadowedValuesInModulesBS from './ShadowedValuesInModules.bs';
+
+export const Nested_M_a: number = ShadowedValuesInModulesBS.Nested[1][0];
+
+export const Nested_num: number = ShadowedValuesInModulesBS.Nested[4];