+++ Changes.md
@@ -1,5 +1,6 @@
 # master
 - Avoid underscore when emitting conversion function for wrapJsForReason.
+- Support inner modules with shadowed values.
 
 # 2.15.0
 - Omit underscore as function name when emitting conversions.